Bounties clean up - separate event bounties into folders
