+++ src/main/java/com/ebicep/warlords/pve/bountysystem/bounties/gardenofhesperides/AcropolisSlayerI.java
@@ -1,55 +1,55 @@
-package com.ebicep.warlords.pve.bountysystem.bounties;
+package com.ebicep.warlords.pve.bountysystem.bounties.gardenofhesperides;
 
 import com.ebicep.warlords.database.repositories.events.pojos.DatabaseGameEvent;
 import com.ebicep.warlords.events.player.ingame.WarlordsDeathEvent;
 import com.ebicep.warlords.game.Game;
-import com.ebicep.warlords.game.option.pve.wavedefense.events.modes.TartarusOption;
+import com.ebicep.warlords.game.option.pve.wavedefense.events.modes.TheAcropolisOption;
 import com.ebicep.warlords.player.ingame.WarlordsNPC;
 import com.ebicep.warlords.pve.bountysystem.AbstractBounty;
 import com.ebicep.warlords.pve.bountysystem.Bounty;
 import com.ebicep.warlords.pve.bountysystem.costs.EventCost;
-import com.ebicep.warlords.pve.bountysystem.rewards.events.GardenOfHesperides2;
+import com.ebicep.warlords.pve.bountysystem.rewards.events.GardenOfHesperides1;
 import com.ebicep.warlords.pve.bountysystem.trackers.TracksDuringGame;
-import com.ebicep.warlords.pve.mobs.events.gardenofhesperides.God;
+import com.ebicep.warlords.pve.mobs.events.gardenofhesperides.LesserGod;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.springframework.data.annotation.Transient;
 
-public class TartarusSlayerI extends AbstractBounty implements TracksDuringGame, EventCost, GardenOfHesperides2 {
+public class AcropolisSlayerI extends AbstractBounty implements TracksDuringGame, EventCost, GardenOfHesperides1 {
 
     @Transient
     private int newKills = 0;
 
     @Override
     public String getName() {
-        return "Tartarus Slayer";
+        return "Acropolis Slayer";
     }
 
     @Override
     public String getDescription() {
-        return "Defeat the Gods of Tartarus 10 times.";
+        return "Defeat 3 of the Lesser Gods in the Acropolis.";
     }
 
     @Override
     public int getTarget() {
-        return 10;
+        return 3;
     }
 
     @Override
     public Bounty getBounty() {
-        return Bounty.TARTARUS_SLAYER_I;
+        return Bounty.ACROPOLIS_SLAYER_I;
     }
 
     @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
     public void onKill(WarlordsDeathEvent event) {
-        if (event.getWarlordsEntity() instanceof WarlordsNPC warlordsNPC && warlordsNPC.getMob() instanceof God) {
+        if (event.getWarlordsEntity() instanceof WarlordsNPC warlordsNPC && warlordsNPC.getMob() instanceof LesserGod) {
             newKills++;
         }
     }
 
     @Override
     public boolean trackGame(Game game) {
-        return DatabaseGameEvent.eventIsActive() && game.getOptions().stream().anyMatch(option -> option instanceof TartarusOption);
+        return DatabaseGameEvent.eventIsActive() && game.getOptions().stream().anyMatch(option -> option instanceof TheAcropolisOption);
     }
 
     @Override
@@ -61,5 +61,4 @@ public class TartarusSlayerI extends AbstractBounty implements TracksDuringGame,
     public long getNewValue() {
         return newKills;
     }
-
 }
