+++ src/main/java/com/ebicep/warlords/pve/bountysystem/bounties/gardenofhesperides/OrderOfThingsI.java
@@ -1,4 +1,4 @@
-package com.ebicep.warlords.pve.bountysystem.bounties;
+package com.ebicep.warlords.pve.bountysystem.bounties.gardenofhesperides;
 
 import com.ebicep.warlords.database.repositories.events.pojos.DatabaseGameEvent;
 import com.ebicep.warlords.events.player.ingame.WarlordsDeathEvent;
@@ -10,7 +10,6 @@ import com.ebicep.warlords.pve.bountysystem.Bounty;
 import com.ebicep.warlords.pve.bountysystem.costs.EventCost;
 import com.ebicep.warlords.pve.bountysystem.rewards.events.GardenOfHesperides2;
 import com.ebicep.warlords.pve.bountysystem.trackers.TracksDuringGame;
-import com.ebicep.warlords.pve.mobs.Mob;
 import com.ebicep.warlords.pve.mobs.events.gardenofhesperides.EventHades;
 import com.ebicep.warlords.pve.mobs.events.gardenofhesperides.EventPoseidon;
 import com.ebicep.warlords.pve.mobs.events.gardenofhesperides.EventZeus;
