+++ src/main/java/com/ebicep/warlords/pve/bountysystem/bounties/gardenofhesperides/StateOfMindI.java
@@ -1,4 +1,4 @@
-package com.ebicep.warlords.pve.bountysystem.bounties;
+package com.ebicep.warlords.pve.bountysystem.bounties.gardenofhesperides;
 
 import com.ebicep.warlords.database.repositories.events.pojos.DatabaseGameEvent;
 import com.ebicep.warlords.events.player.ingame.WarlordsDeathEvent;
@@ -10,39 +10,47 @@ import com.ebicep.warlords.pve.bountysystem.Bounty;
 import com.ebicep.warlords.pve.bountysystem.costs.EventCost;
 import com.ebicep.warlords.pve.bountysystem.rewards.events.GardenOfHesperides1;
 import com.ebicep.warlords.pve.bountysystem.trackers.TracksDuringGame;
-import com.ebicep.warlords.pve.mobs.events.gardenofhesperides.Teras;
+import com.ebicep.warlords.pve.mobs.events.gardenofhesperides.EventCronus;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.springframework.data.annotation.Transient;
 
-public class TerasTormentI extends AbstractBounty implements TracksDuringGame, EventCost, GardenOfHesperides1 {
+public class StateOfMindI extends AbstractBounty implements TracksDuringGame, EventCost, GardenOfHesperides1 {
 
     @Transient
     private int newKills = 0;
 
     @Override
     public String getName() {
-        return "Teras Torment";
+        return "State of Mind";
     }
 
     @Override
     public String getDescription() {
-        return "Defeat 100 Teras Mobs in the Acropolis.";
+        return "Defeat Cronus while he is rejuvenating.";
     }
 
     @Override
     public int getTarget() {
-        return 100;
+        return 1;
     }
 
     @Override
     public Bounty getBounty() {
-        return Bounty.TERAS_TORMENT_I;
+        return Bounty.STATE_OF_MIND_I;
+    }
+
+    @Override
+    public void reset() {
+        newKills = 0;
     }
 
     @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
     public void onKill(WarlordsDeathEvent event) {
-        if (event.getWarlordsEntity() instanceof WarlordsNPC warlordsNPC && warlordsNPC.getMob() instanceof Teras) {
+        if (event.getWarlordsEntity() instanceof WarlordsNPC warlordsNPC &&
+                warlordsNPC.getMob() instanceof EventCronus cronus &&
+                cronus.diedDuringRejuvenate()
+        ) {
             newKills++;
         }
     }
@@ -50,11 +58,6 @@ public class TerasTormentI extends AbstractBounty implements TracksDuringGame, E
     @Override
     public boolean trackGame(Game game) {
         return DatabaseGameEvent.eventIsActive() && game.getOptions().stream().anyMatch(option -> option instanceof TheAcropolisOption);
-    }
-
-    @Override
-    public void reset() {
-        newKills = 0;
     }
 
     @Override
