+++ src/main/java/com/ebicep/warlords/pve/bountysystem/bounties/gardenofhesperides/TartarusFlawlessI.java
@@ -1,10 +1,8 @@
-package com.ebicep.warlords.pve.bountysystem.bounties;
+package com.ebicep.warlords.pve.bountysystem.bounties.gardenofhesperides;
 
 import com.ebicep.warlords.database.repositories.events.pojos.DatabaseGameEvent;
 import com.ebicep.warlords.events.game.WarlordsGameTriggerWinEvent;
 import com.ebicep.warlords.game.Game;
-import com.ebicep.warlords.game.option.RecordTimeElapsedOption;
-import com.ebicep.warlords.game.option.pve.wavedefense.WinByMaxWaveClearOption;
 import com.ebicep.warlords.game.option.pve.wavedefense.events.modes.TartarusOption;
 import com.ebicep.warlords.player.ingame.WarlordsPlayer;
 import com.ebicep.warlords.pve.bountysystem.AbstractBounty;
@@ -14,16 +12,16 @@ import com.ebicep.warlords.pve.bountysystem.costs.EventCost;
 import com.ebicep.warlords.pve.bountysystem.rewards.events.GardenOfHesperides2;
 import com.ebicep.warlords.pve.bountysystem.trackers.TracksPostGame;
 
-public class WithinTheTimeI extends AbstractBounty implements TracksPostGame, EventCost, GardenOfHesperides2 {
+public class TartarusFlawlessI extends AbstractBounty implements TracksPostGame, EventCost, GardenOfHesperides2 {
 
     @Override
     public String getName() {
-        return "Within the Time";
+        return "Tartarus Flawless";
     }
 
     @Override
     public String getDescription() {
-        return "Complete Tartarus within 10 minutes.";
+        return "Complete Tartarus without dying.";
     }
 
     @Override
@@ -33,7 +31,7 @@ public class WithinTheTimeI extends AbstractBounty implements TracksPostGame, Ev
 
     @Override
     public Bounty getBounty() {
-        return Bounty.WITHIN_THE_TIME_I;
+        return Bounty.TARTARUS_FLAWLESS_I;
     }
 
     @Override
@@ -41,15 +39,12 @@ public class WithinTheTimeI extends AbstractBounty implements TracksPostGame, Ev
         if (!DatabaseGameEvent.eventIsActive()) {
             return;
         }
-        if (BountyUtils.getOptionFromGame(game, TartarusOption.class).isEmpty()) {
-            return;
-        }
-        BountyUtils.getOptionFromGame(game, RecordTimeElapsedOption.class)
-                   .ifPresent(recordTimeElapsedOption -> {
-                       if (gameWinEvent.getCause() instanceof WinByMaxWaveClearOption && recordTimeElapsedOption.getTicksElapsed() < 10 * 60 * 20) {
-                           value++;
-                       }
-                   });
+        BountyUtils.getOptionFromGame(game, TartarusOption.class).ifPresent(acropolisOption -> {
+            int deaths = warlordsPlayer.getMinuteStats().total().getDeaths();
+            if (deaths == 0) {
+                value++;
+            }
+        });
     }
 
 }
