+++ src/main/java/com/ebicep/warlords/pve/bountysystem/bounties/gardenofhesperides/TartarusSlayerI.java
@@ -1,55 +1,55 @@
-package com.ebicep.warlords.pve.bountysystem.bounties;
+package com.ebicep.warlords.pve.bountysystem.bounties.gardenofhesperides;
 
 import com.ebicep.warlords.database.repositories.events.pojos.DatabaseGameEvent;
 import com.ebicep.warlords.events.player.ingame.WarlordsDeathEvent;
 import com.ebicep.warlords.game.Game;
-import com.ebicep.warlords.game.option.pve.wavedefense.events.modes.GrimoiresGraveyardOption;
+import com.ebicep.warlords.game.option.pve.wavedefense.events.modes.TartarusOption;
 import com.ebicep.warlords.player.ingame.WarlordsNPC;
 import com.ebicep.warlords.pve.bountysystem.AbstractBounty;
 import com.ebicep.warlords.pve.bountysystem.Bounty;
 import com.ebicep.warlords.pve.bountysystem.costs.EventCost;
-import com.ebicep.warlords.pve.bountysystem.rewards.events.LibraryArchives1;
+import com.ebicep.warlords.pve.bountysystem.rewards.events.GardenOfHesperides2;
 import com.ebicep.warlords.pve.bountysystem.trackers.TracksDuringGame;
-import com.ebicep.warlords.pve.mobs.events.libraryarchives.EventTheArchivist;
+import com.ebicep.warlords.pve.mobs.events.gardenofhesperides.God;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.springframework.data.annotation.Transient;
 
-public class ArchivistHunterI extends AbstractBounty implements TracksDuringGame, EventCost, LibraryArchives1 {
+public class TartarusSlayerI extends AbstractBounty implements TracksDuringGame, EventCost, GardenOfHesperides2 {
 
     @Transient
     private int newKills = 0;
 
     @Override
     public String getName() {
-        return "Archivist Hunter";
+        return "Tartarus Slayer";
     }
 
     @Override
     public String getDescription() {
-        return "Defeat The Archivist " + getTarget() + " times.";
+        return "Defeat the Gods of Tartarus 10 times.";
     }
 
     @Override
     public int getTarget() {
-        return 20;
+        return 10;
     }
 
     @Override
     public Bounty getBounty() {
-        return Bounty.ARCHIVIST_HUNTER_I;
+        return Bounty.TARTARUS_SLAYER_I;
     }
 
     @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
     public void onKill(WarlordsDeathEvent event) {
-        if (event.getWarlordsEntity() instanceof WarlordsNPC warlordsNPC && warlordsNPC.getMob() instanceof EventTheArchivist) {
+        if (event.getWarlordsEntity() instanceof WarlordsNPC warlordsNPC && warlordsNPC.getMob() instanceof God) {
             newKills++;
         }
     }
 
     @Override
     public boolean trackGame(Game game) {
-        return DatabaseGameEvent.eventIsActive() && game.getOptions().stream().anyMatch(option -> option instanceof GrimoiresGraveyardOption);
+        return DatabaseGameEvent.eventIsActive() && game.getOptions().stream().anyMatch(option -> option instanceof TartarusOption);
     }
 
     @Override
@@ -61,4 +61,5 @@ public class ArchivistHunterI extends AbstractBounty implements TracksDuringGame
     public long getNewValue() {
         return newKills;
     }
+
 }
