+++ src/main/java/com/ebicep/warlords/pve/bountysystem/bounties/gardenofhesperides/TerasTormentI.java
@@ -1,4 +1,4 @@
-package com.ebicep.warlords.pve.bountysystem.bounties;
+package com.ebicep.warlords.pve.bountysystem.bounties.gardenofhesperides;
 
 import com.ebicep.warlords.database.repositories.events.pojos.DatabaseGameEvent;
 import com.ebicep.warlords.events.player.ingame.WarlordsDeathEvent;
@@ -10,39 +10,39 @@ import com.ebicep.warlords.pve.bountysystem.Bounty;
 import com.ebicep.warlords.pve.bountysystem.costs.EventCost;
 import com.ebicep.warlords.pve.bountysystem.rewards.events.GardenOfHesperides1;
 import com.ebicep.warlords.pve.bountysystem.trackers.TracksDuringGame;
-import com.ebicep.warlords.pve.mobs.events.gardenofhesperides.LesserGod;
+import com.ebicep.warlords.pve.mobs.events.gardenofhesperides.Teras;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.springframework.data.annotation.Transient;
 
-public class AcropolisSlayerI extends AbstractBounty implements TracksDuringGame, EventCost, GardenOfHesperides1 {
+public class TerasTormentI extends AbstractBounty implements TracksDuringGame, EventCost, GardenOfHesperides1 {
 
     @Transient
     private int newKills = 0;
 
     @Override
     public String getName() {
-        return "Acropolis Slayer";
+        return "Teras Torment";
     }
 
     @Override
     public String getDescription() {
-        return "Defeat 3 of the Lesser Gods in the Acropolis.";
+        return "Defeat 100 Teras Mobs in the Acropolis.";
     }
 
     @Override
     public int getTarget() {
-        return 3;
+        return 100;
     }
 
     @Override
     public Bounty getBounty() {
-        return Bounty.ACROPOLIS_SLAYER_I;
+        return Bounty.TERAS_TORMENT_I;
     }
 
     @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
     public void onKill(WarlordsDeathEvent event) {
-        if (event.getWarlordsEntity() instanceof WarlordsNPC warlordsNPC && warlordsNPC.getMob() instanceof LesserGod) {
+        if (event.getWarlordsEntity() instanceof WarlordsNPC warlordsNPC && warlordsNPC.getMob() instanceof Teras) {
             newKills++;
         }
     }
