+++ src/main/java/com/ebicep/warlords/pve/bountysystem/bounties/libraryarchives/CodexChaosI.java
@@ -1,7 +1,6 @@
-package com.ebicep.warlords.pve.bountysystem.bounties;
+package com.ebicep.warlords.pve.bountysystem.bounties.libraryarchives;
 
 import com.ebicep.warlords.database.repositories.events.pojos.DatabaseGameEvent;
-import com.ebicep.warlords.events.player.ingame.WarlordsDeathEvent;
 import com.ebicep.warlords.game.Game;
 import com.ebicep.warlords.game.option.pve.wavedefense.events.modes.ForgottenCodexOption;
 import com.ebicep.warlords.game.option.pve.wavedefense.events.modes.GrimoiresGraveyardOption;
@@ -25,13 +24,23 @@ import org.springframework.data.mongodb.core.mapping.Field;
 import java.util.Arrays;
 import java.util.List;
 
-public class ForgottenSlayerI extends AbstractBounty implements TracksDuringGame, EventCost, LibraryArchives2 {
 
-    @Field("inquisiteurs_defeated")
-    private int[] inquisiteursDefeated = {0, 0, 0}; // index > 0 = EGA, 1 = EWA, 2 = VPA | element value > times killed
+public class CodexChaosI extends AbstractBounty implements TracksDuringGame, EventCost, LibraryArchives2 {
+
+    @Field("killing_blows_experienced")
+    private int[] killingBlowsExperienced = {0, 0, 0};  // index > 0 = EGA, 1 = EWA, 2 = VPA | element value > # of blows
     @Transient
-    private int[] newInquisiteursDefeated = {0, 0, 0};
+    private int[] newKillingBlowsExperienced = {0, 0, 0};
 
+    @Override
+    public String getName() {
+        return "Codex Chaos";
+    }
+
+    @Override
+    public String getDescription() {
+        return "Experience the Killing Blow ability from each Inquisiteur variant 3 times.";
+    }
 
     @Nullable
     @Override
@@ -42,31 +51,21 @@ public class ForgottenSlayerI extends AbstractBounty implements TracksDuringGame
         return List.of(
                 ComponentBuilder.create("Progress: ", NamedTextColor.GRAY).build(),
                 ComponentBuilder.create("  EGA: ", NamedTextColor.GRAY)
-                                .text(NumberFormat.addCommaAndRound(inquisiteursDefeated[0]), NamedTextColor.GOLD)
+                                .text(NumberFormat.addCommaAndRound(killingBlowsExperienced[0]), NamedTextColor.GOLD)
                                 .text("/", NamedTextColor.AQUA)
-                                .text("5", NamedTextColor.GOLD)
+                                .text("3", NamedTextColor.GOLD)
                                 .build(),
                 ComponentBuilder.create("  EWA: ", NamedTextColor.GRAY)
-                                .text(NumberFormat.addCommaAndRound(inquisiteursDefeated[1]), NamedTextColor.GOLD)
+                                .text(NumberFormat.addCommaAndRound(killingBlowsExperienced[1]), NamedTextColor.GOLD)
                                 .text("/", NamedTextColor.AQUA)
-                                .text("5", NamedTextColor.GOLD)
+                                .text("3", NamedTextColor.GOLD)
                                 .build(),
                 ComponentBuilder.create("  VPA: ", NamedTextColor.GRAY)
-                                .text(NumberFormat.addCommaAndRound(inquisiteursDefeated[2]), NamedTextColor.GOLD)
+                                .text(NumberFormat.addCommaAndRound(killingBlowsExperienced[2]), NamedTextColor.GOLD)
                                 .text("/", NamedTextColor.AQUA)
-                                .text("5", NamedTextColor.GOLD)
+                                .text("3", NamedTextColor.GOLD)
                                 .build()
         );
-    }
-
-    @Override
-    public String getName() {
-        return "Forgotten Slayer";
-    }
-
-    @Override
-    public String getDescription() {
-        return "Defeat each of the Inquisiteurs 5 times.";
     }
 
     @Override
@@ -76,13 +75,13 @@ public class ForgottenSlayerI extends AbstractBounty implements TracksDuringGame
 
     @Override
     public Bounty getBounty() {
-        return Bounty.FORGOTTEN_SLAYER_I;
+        return Bounty.CODEX_CHAOS_I;
     }
 
     @Override
     public void reset() {
         for (int i = 0; i < 3; i++) {
-            inquisiteursDefeated[i] = 0;
+            killingBlowsExperienced[i] = 0;
         }
     }
 
@@ -96,23 +95,23 @@ public class ForgottenSlayerI extends AbstractBounty implements TracksDuringGame
     @Override
     public void apply(AbstractBounty bounty) {
         for (int i = 0; i < 3; i++) {
-            inquisiteursDefeated[i] += newInquisiteursDefeated[i];
+            killingBlowsExperienced[i] += newKillingBlowsExperienced[i];
         }
         TracksDuringGame.super.apply(bounty);
     }
 
     @Override
     public long getNewValue() {
-        return Arrays.stream(inquisiteursDefeated).allMatch(integer -> integer >= 5) ? 1 : 0;
+        return Arrays.stream(killingBlowsExperienced).allMatch(integer -> integer >= 3) ? 1 : 0;
     }
 
     @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
-    public void onKill(WarlordsDeathEvent event) {
-        if (event.getWarlordsEntity() instanceof WarlordsNPC warlordsNPC && warlordsNPC.getMob() instanceof EventInquisiteur) {
+    public void onKillingBlow(EventInquisiteur.EventInquisteurKillingBlowEvent event) {
+        if (event.getWarlordsEntity() instanceof WarlordsNPC warlordsNPC) {
             switch (warlordsNPC.getMob().getMobRegistry()) {
-                case EVENT_INQUISITEUR_EGA -> newInquisiteursDefeated[0]++;
-                case EVENT_INQUISITEUR_EWA -> newInquisiteursDefeated[1]++;
-                case EVENT_INQUISITEUR_VPA -> newInquisiteursDefeated[2]++;
+                case EVENT_INQUISITEUR_EGA -> newKillingBlowsExperienced[0]++;
+                case EVENT_INQUISITEUR_EWA -> newKillingBlowsExperienced[1]++;
+                case EVENT_INQUISITEUR_VPA -> newKillingBlowsExperienced[2]++;
             }
         }
     }
