+++ src/main/java/com/ebicep/warlords/pve/bountysystem/bounties/libraryarchives/CodexContinuedI.java
@@ -1,4 +1,4 @@
-package com.ebicep.warlords.pve.bountysystem.bounties;
+package com.ebicep.warlords.pve.bountysystem.bounties.libraryarchives;
 
 import com.ebicep.warlords.database.repositories.events.pojos.DatabaseGameEvent;
 import com.ebicep.warlords.events.game.WarlordsGameTriggerWinEvent;
