+++ src/main/java/com/ebicep/warlords/pve/bountysystem/bounties/libraryarchives/ForgottenFlawlessI.java
@@ -1,4 +1,4 @@
-package com.ebicep.warlords.pve.bountysystem.bounties;
+package com.ebicep.warlords.pve.bountysystem.bounties.libraryarchives;
 
 import com.ebicep.warlords.database.repositories.events.pojos.DatabaseGameEvent;
 import com.ebicep.warlords.events.game.WarlordsGameTriggerWinEvent;
@@ -11,19 +11,17 @@ import com.ebicep.warlords.pve.bountysystem.BountyUtils;
 import com.ebicep.warlords.pve.bountysystem.costs.EventCost;
 import com.ebicep.warlords.pve.bountysystem.rewards.events.LibraryArchives2;
 import com.ebicep.warlords.pve.bountysystem.trackers.TracksPostGame;
-import com.ebicep.warlords.pve.weapons.AbstractWeapon;
-import com.ebicep.warlords.pve.weapons.weapontypes.legendaries.titles.LibraryArchivesTitle;
 
-public class TakeMyTitleII extends AbstractBounty implements TracksPostGame, EventCost, LibraryArchives2 {
+public class ForgottenFlawlessI extends AbstractBounty implements TracksPostGame, EventCost, LibraryArchives2 {
 
     @Override
     public String getName() {
-        return "Take My Title";
+        return "Forgotten Flawless";
     }
 
     @Override
     public String getDescription() {
-        return "Complete Forgotten Codex with a legendary weapon equipped with a Library Archives title " + getTarget() + " times.";
+        return "Complete Forgotten Codex " + getTarget() + " times without dying.";
     }
 
     @Override
@@ -33,7 +31,7 @@ public class TakeMyTitleII extends AbstractBounty implements TracksPostGame, Eve
 
     @Override
     public Bounty getBounty() {
-        return Bounty.TAKE_MY_TITLE_II;
+        return Bounty.FORGOTTEN_FLAWLESS_I;
     }
 
     @Override
@@ -44,17 +42,11 @@ public class TakeMyTitleII extends AbstractBounty implements TracksPostGame, Eve
         if (BountyUtils.lostGame(gameWinEvent)) {
             return;
         }
-        if (BountyUtils.getOptionFromGame(game, ForgottenCodexOption.class).isEmpty()) {
-            return;
-        }
-        AbstractWeapon weapon = warlordsPlayer.getWeapon();
-        if (weapon == null) {
-            return;
-        }
-        if (!(weapon instanceof LibraryArchivesTitle)) {
-            return;
-        }
-        value++;
+        BountyUtils.getOptionFromGame(game, ForgottenCodexOption.class).ifPresent(option -> {
+            if (warlordsPlayer.getMinuteStats().total().getDeaths() == 0) {
+                value++;
+            }
+        });
     }
 
 }
