+++ src/main/java/com/ebicep/warlords/pve/bountysystem/bounties/libraryarchives/GraveyardFlawlessI.java
@@ -1,37 +1,37 @@
-package com.ebicep.warlords.pve.bountysystem.bounties;
+package com.ebicep.warlords.pve.bountysystem.bounties.libraryarchives;
 
 import com.ebicep.warlords.database.repositories.events.pojos.DatabaseGameEvent;
 import com.ebicep.warlords.events.game.WarlordsGameTriggerWinEvent;
 import com.ebicep.warlords.game.Game;
-import com.ebicep.warlords.game.option.pve.wavedefense.events.modes.TheAcropolisOption;
+import com.ebicep.warlords.game.option.pve.wavedefense.events.modes.GrimoiresGraveyardOption;
 import com.ebicep.warlords.player.ingame.WarlordsPlayer;
 import com.ebicep.warlords.pve.bountysystem.AbstractBounty;
 import com.ebicep.warlords.pve.bountysystem.Bounty;
 import com.ebicep.warlords.pve.bountysystem.BountyUtils;
 import com.ebicep.warlords.pve.bountysystem.costs.EventCost;
-import com.ebicep.warlords.pve.bountysystem.rewards.events.GardenOfHesperides1;
+import com.ebicep.warlords.pve.bountysystem.rewards.events.LibraryArchives1;
 import com.ebicep.warlords.pve.bountysystem.trackers.TracksPostGame;
 
-public class AcropolisFlawlessI extends AbstractBounty implements TracksPostGame, EventCost, GardenOfHesperides1 {
+public class GraveyardFlawlessI extends AbstractBounty implements TracksPostGame, EventCost, LibraryArchives1 {
 
     @Override
     public String getName() {
-        return "Acropolis Flawless";
+        return "Graveyard Flawless";
     }
 
     @Override
     public String getDescription() {
-        return "Complete Acropolis without dying.";
+        return "Complete Grimoire’s Graveyard " + getTarget() + " times without dying.";
     }
 
     @Override
     public int getTarget() {
-        return 1;
+        return 3;
     }
 
     @Override
     public Bounty getBounty() {
-        return Bounty.ACROPOLIS_FLAWLESS_I;
+        return Bounty.GRAVEYARD_FLAWLESS_I;
     }
 
     @Override
@@ -39,11 +39,14 @@ public class AcropolisFlawlessI extends AbstractBounty implements TracksPostGame
         if (!DatabaseGameEvent.eventIsActive()) {
             return;
         }
-        BountyUtils.getOptionFromGame(game, TheAcropolisOption.class).ifPresent(acropolisOption -> {
-            int deaths = warlordsPlayer.getMinuteStats().total().getDeaths();
-            if (deaths == 0) {
+        if (BountyUtils.lostGame(gameWinEvent)) {
+            return;
+        }
+        BountyUtils.getOptionFromGame(game, GrimoiresGraveyardOption.class).ifPresent(option -> {
+            if (warlordsPlayer.getMinuteStats().total().getDeaths() == 0) {
                 value++;
             }
         });
     }
+
 }
