+++ src/main/java/com/ebicep/warlords/pve/bountysystem/bounties/libraryarchives/GrimoiresGriefI.java
@@ -1,4 +1,4 @@
-package com.ebicep.warlords.pve.bountysystem.bounties;
+package com.ebicep.warlords.pve.bountysystem.bounties.libraryarchives;
 
 import com.ebicep.warlords.database.repositories.events.pojos.DatabaseGameEvent;
 import com.ebicep.warlords.events.player.ingame.WarlordsDamageHealingFinalEvent;
@@ -11,41 +11,54 @@ import com.ebicep.warlords.pve.bountysystem.Bounty;
 import com.ebicep.warlords.pve.bountysystem.costs.EventCost;
 import com.ebicep.warlords.pve.bountysystem.rewards.events.LibraryArchives1;
 import com.ebicep.warlords.pve.bountysystem.trackers.TracksDuringGame;
-import com.ebicep.warlords.pve.mobs.events.libraryarchives.EventNecronomiconGrimoire;
+import com.ebicep.warlords.pve.mobs.events.libraryarchives.EventGrimoire;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.springframework.data.annotation.Transient;
 
 import java.util.Objects;
 
-public class ThatWasCloseI extends AbstractBounty implements TracksDuringGame, EventCost, LibraryArchives1 {
+public class GrimoiresGriefI extends AbstractBounty implements TracksDuringGame, EventCost, LibraryArchives1 {
 
     @Transient
-    private int newValue = 0;
+    private int newKills = 0;
 
     @Override
     public String getName() {
-        return "That Was Close";
+        return "Grimoire's Grief";
     }
 
     @Override
     public String getDescription() {
-        return "Survive the smite of a Necronomicon " + getTarget() + " times.";
+        return "Kill " + getTarget() + " Boss Minion Grimoire.";
     }
 
     @Override
     public int getTarget() {
-        return 15;
+        return 100;
     }
 
     @Override
     public Bounty getBounty() {
-        return Bounty.THAT_WAS_CLOSE_I;
+        return Bounty.GRIMOIRES_GRIEF_I;
     }
 
     @Override
     public void reset() {
-        newValue = 0;
+        newKills = 0;
+    }
+
+    @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
+    public void onFinalDamageHeal(WarlordsDamageHealingFinalEvent event) {
+        if (!Objects.equals(event.getSource().getUuid(), uuid)) {
+            return;
+        }
+        if (!event.isDead()) {
+            return;
+        }
+        if (event.getWarlordsEntity() instanceof WarlordsNPC warlordsNPC && warlordsNPC.getMob() instanceof EventGrimoire) {
+            newKills++;
+        }
     }
 
     @Override
@@ -57,23 +70,6 @@ public class ThatWasCloseI extends AbstractBounty implements TracksDuringGame, E
 
     @Override
     public long getNewValue() {
-        return newValue;
-    }
-
-    @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
-    public void onFinalDamageHeal(WarlordsDamageHealingFinalEvent event) {
-        if (!Objects.equals(event.getWarlordsEntity().getUuid(), uuid)) {
-            return;
-        }
-        if (!(event.getSource() instanceof WarlordsNPC warlordsNPC && warlordsNPC.getMob() instanceof EventNecronomiconGrimoire)) {
-            return;
-        }
-        if (!event.getCause().equals("Smite")) {
-            return;
-        }
-        if (event.isDead()) {
-            return;
-        }
-        newValue++;
+        return newKills;
     }
 }
