+++ src/main/java/com/ebicep/warlords/pve/bountysystem/bounties/libraryarchives/TakeMyTitleII.java
@@ -1,27 +1,29 @@
-package com.ebicep.warlords.pve.bountysystem.bounties;
+package com.ebicep.warlords.pve.bountysystem.bounties.libraryarchives;
 
 import com.ebicep.warlords.database.repositories.events.pojos.DatabaseGameEvent;
 import com.ebicep.warlords.events.game.WarlordsGameTriggerWinEvent;
 import com.ebicep.warlords.game.Game;
-import com.ebicep.warlords.game.option.pve.wavedefense.events.modes.GrimoiresGraveyardOption;
+import com.ebicep.warlords.game.option.pve.wavedefense.events.modes.ForgottenCodexOption;
 import com.ebicep.warlords.player.ingame.WarlordsPlayer;
 import com.ebicep.warlords.pve.bountysystem.AbstractBounty;
 import com.ebicep.warlords.pve.bountysystem.Bounty;
 import com.ebicep.warlords.pve.bountysystem.BountyUtils;
 import com.ebicep.warlords.pve.bountysystem.costs.EventCost;
-import com.ebicep.warlords.pve.bountysystem.rewards.events.LibraryArchives1;
+import com.ebicep.warlords.pve.bountysystem.rewards.events.LibraryArchives2;
 import com.ebicep.warlords.pve.bountysystem.trackers.TracksPostGame;
+import com.ebicep.warlords.pve.weapons.AbstractWeapon;
+import com.ebicep.warlords.pve.weapons.weapontypes.legendaries.titles.LibraryArchivesTitle;
 
-public class GraveyardFlawlessI extends AbstractBounty implements TracksPostGame, EventCost, LibraryArchives1 {
+public class TakeMyTitleII extends AbstractBounty implements TracksPostGame, EventCost, LibraryArchives2 {
 
     @Override
     public String getName() {
-        return "Graveyard Flawless";
+        return "Take My Title";
     }
 
     @Override
     public String getDescription() {
-        return "Complete Grimoire’s Graveyard " + getTarget() + " times without dying.";
+        return "Complete Forgotten Codex with a legendary weapon equipped with a Library Archives title " + getTarget() + " times.";
     }
 
     @Override
@@ -31,7 +33,7 @@ public class GraveyardFlawlessI extends AbstractBounty implements TracksPostGame
 
     @Override
     public Bounty getBounty() {
-        return Bounty.GRAVEYARD_FLAWLESS_I;
+        return Bounty.TAKE_MY_TITLE_II;
     }
 
     @Override
@@ -42,11 +44,17 @@ public class GraveyardFlawlessI extends AbstractBounty implements TracksPostGame
         if (BountyUtils.lostGame(gameWinEvent)) {
             return;
         }
-        BountyUtils.getOptionFromGame(game, GrimoiresGraveyardOption.class).ifPresent(option -> {
-            if (warlordsPlayer.getMinuteStats().total().getDeaths() == 0) {
-                value++;
-            }
-        });
+        if (BountyUtils.getOptionFromGame(game, ForgottenCodexOption.class).isEmpty()) {
+            return;
+        }
+        AbstractWeapon weapon = warlordsPlayer.getWeapon();
+        if (weapon == null) {
+            return;
+        }
+        if (!(weapon instanceof LibraryArchivesTitle)) {
+            return;
+        }
+        value++;
     }
 
 }
